fix(#1910): satisfy structural ratchets by extracting bounded-state code to siblings
The bounded-JSONL-state work grew watchlist/grandfather god-files
(orchestrator.ts, storage.ts, config.ts, types.ts) and added a direct
storage.ts import, tripping the #1520/#1995/#1533 ratchets. Fix without
baseline changes or behavior changes:

- New bounded-jsonl-state.ts: BoundedJsonlStateConfig interface (PluginConfig
  now extends it) + parseBoundedJsonlStateConfig; config.ts spreads it,
  types.ts extends it.
- New offline-sync-exclude-globs.ts: DEFAULT_OFFLINE_SYNC_EXCLUDE_GLOBS moved
  out of offline-sync.ts.
- New storage/memory-lifecycle-ledger-access.ts: readAll/readBounded lifecycle
  ledger readers extracted from storage.ts; storage.ts keeps thin delegators.
- maintenance.ts imports the StorageManager type via ../index.js (matching the
  existing coordinator convention) instead of ../storage.js, so
  directStorageImports returns to 42.
- orchestrator.ts: drop the redundant getQmd comment (duplicated by the
  maintenance deps doc) and the redundant #1910 comment, and fold the two
  storage deps onto one line to offset the new constructor wiring at the file's
  pinned ceiling.

Ratchets green (no baseline edit); core tsc --noEmit clean; 275 focused
bounded-state tests pass.

Part of #1910
