+++ packages/remnic-core/src/orchestrator.ts
@@ -1420,18 +1420,10 @@ export class Orchestrator {
     this.qmd = createSearchBackend(config);
     this.maintenanceScheduler = new MaintenanceScheduler({
       config,
-      // Live accessor: the orchestrator reassigns this.qmd to NoopSearchBackend
-      // after construction when the collection is missing (initialize /
-      // startupSearchSync); the scheduler must read the current backend so
-      // debounced maintenance never runs against a stale/disposed one.
       getQmd: () => this.qmd,
       namespaceSearchRouter: this.namespaceSearchRouter,
       namespaceCatalog: this.namespaceCatalog,
-      // Issue #1910: thread the live secure-store context so lifecycle-ledger
-      // auto-compaction rebuilds encrypted ledgers with the active key, and
-      // reaches per-namespace ledgers via the router's secure-aware storages.
-      getStorage: () => this.storage,
-      storageForNamespace: (namespace) => this.storageRouter.storageFor(namespace),
+      getStorage: () => this.storage, storageForNamespace: (namespace) => this.storageRouter.storageFor(namespace),
     });
     // Issue #1526: background extraction queue lives on its own coordinator.
     this.extractionQueueCoordinator = new ExtractionQueueCoordinator();
